test: make client test independet from the api (#302)
Our client_test is not a unit test; we rely on the API to set the provider regularly, and within the code, we rely on the singleton. With these changes, we mock the API interface and provide it to the client's existing factory method. So each of the tests is independent from the others.

There are still some tests, which i would rather declare as e2e or integration tests, which i did not tackle, or where i do think it makes sense to keep them as is.

---------

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>
Co-authored-by: Bernd Warmuth <bernd.warmuth@dynatrace.com>