refactor: remove dependency on golang.org/x/exp
The exp module does not have the same compatibility guarantees as the
standard library. The fact that it stays on a pseudo-version instead of
a proper semver tag, and the possibility for backwards-incompatible
changes, makes depending on exp in a widely-imported library a bit
risky. If a dependent of the SDK updated the version of exp they relied
on, the MVS algorithm could select a version that broke the SDK's usage
of exp's packages.

Out of an abundance of caution, I'm trying to reduce the use of exp in
our private shared libraries for this reason, and the OpenFeature SDK is
our only external dependency that depends on it it. The uses are pretty
small. I've replaced them with existing standard library functiosn where
possible, or in one place implemented the functionality myself, since
`maps.Values` is only available in the standard library from go 1.23.

While in the area, I've also updated a couple of functions that are
already replaceable with functions from the standard slices package.

Signed-off-by: Eli Treuherz <et@arenko.group>