fix: Correct custom strategy implementation
While reviewing the code to write additional documentation I noticed that there's a problem with the way custom strategies
are implemented. There's no way for it to process multiple providers! All the other strategy functions are created using
a closure to include the providers. This didn't exist for custom strategies so it would never actually have a provider to call!

I corrected this by adding a new function type `StrategyConstructor` for custom strategies to use to be able to close over
the providers.

Signed-off-by: Jordan Blacker <jblacker@justworks.com>